server wrapped for lambda
