+++ .gitignore
@@ -46,3 +46,5 @@ Thumbs.db
 **/*.env
 
 .nx
+.serverless
+.build
