+++ apps/server/Dockerfile
@@ -1,24 +1,14 @@
-# This file is generated by Nx.
-#
-# Build the docker image with `npx nx docker-build server`.
-# Tip: Modify "docker-build" options in project.json to change docker build args.
-#
-# Run the container with `docker run -p 3000:3000 -t server`.
-FROM docker.io/node:lts-alpine
+FROM public.ecr.aws/lambda/nodejs:22
 
-ENV HOST=0.0.0.0
-ENV PORT=3000
 
-WORKDIR /app
+# Set working directory
+WORKDIR ${LAMBDA_TASK_ROOT}
 
-RUN addgroup --system server && \
-          adduser --system -G server server
+# Copy the build output to the container
+COPY dist/apps/server .
 
-COPY dist/apps/server server/
-RUN chown -R server:server .
+# Install production dependencies
+RUN npm install --omit=dev
 
-# You can remove this install step if you build with `--bundle` option.
-# The bundled output will include external dependencies.
-RUN npm --prefix server --omit=dev -f install
 
-CMD [ "node", "server" ]
+CMD [ "main.handler" ]
