+++ package.json
@@ -28,9 +28,11 @@
     "@types/cookie-parser": "^1.4.7",
     "@types/express": "^5.0.0",
     "@types/passport-jwt": "^4.0.1",
+    "@vendia/serverless-express": "^4.12.6",
+    "aws-lambda": "^1.0.7",
     "aws-sdk": "^2.1692.0",
     "axios": "^1.7.7",
-    "bcrypt": "^5.1.1",
+    "bcryptjs": "^2.4.3",
     "class-transformer": "^0.5.1",
     "class-validator": "^0.14.1",
     "compression": "^1.7.5",
@@ -65,6 +67,7 @@
     "@nestjs/core": "^10.4.8",
     "@nestjs/schematics": "^10.0.1",
     "@nestjs/testing": "^10.4.8",
+    "@nrwl/node": "^19.8.4",
     "@nrwl/workspace": "^19.8.4",
     "@nx/cypress": "19.8.4",
     "@nx/eslint": "20.1.0",
@@ -80,7 +83,9 @@
     "@swc-node/register": "~1.9.1",
     "@swc/core": "~1.5.7",
     "@swc/helpers": "~0.5.11",
+    "@types/aws-lambda": "^8.10.146",
     "@types/bcrypt": "^5.0.2",
+    "@types/bcryptjs": "^2.4.6",
     "@types/compression": "^1.7.5",
     "@types/hpp": "^0.2.6",
     "@types/jest": "^29.5.14",
@@ -100,6 +105,7 @@
     "postcss": "^8.4.49",
     "prettier": "^2.6.2",
     "tailwindcss": "^3.4.15",
+    "terser-webpack-plugin": "^5.3.10",
     "ts-jest": "^29.2.5",
     "ts-node": "10.9.1",
     "tslib": "^2.3.0",
