+++ tsconfig.base.json
@@ -14,6 +14,7 @@
     "skipLibCheck": true,
     "skipDefaultLibCheck": true,
     "baseUrl": ".",
+    "esModuleInterop": true,
     "paths": {
       "@featherly/shared": ["shared/src/index.ts"]
     }
